internal: Add `GeneratorContext` class, allowing for generator re-use
