+++ packages/schemas/src/HypixelApiError.ts
@@ -1,0 +1,9 @@
+export class HypixelApiError extends Error {
+    constructor(cause: string, url: string, options?: ErrorOptions) {
+        super(
+            `${cause}\n` +
+            `URL: ${url}\n`,
+            options
+        );
+    }
+}
