+++ packages/schemas/src/generator-worker.ts
@@ -2,7 +2,7 @@ import workerpool from "workerpool";
 import {JSONSchema4} from "json-schema";
 import chalk from "chalk";
 import toJsonSchema from "gen-json-schema";
-import {logger} from "./tools";
+import {logger} from "./common";
 import Ajv, {Options, ValidateFunction} from "ajv";
 import {diff} from "json-diff";
 import crypto from "crypto";
